refactor: pages
import문 줄이기 및 폴더명 변경
