--- src/page/SignInPage.tsx
@@ -1,8 +1,0 @@
-import React from "react";
-import AuthHOC from "../components/HOC/AuthHOC";
-
-const SignInPage = () => {
-  return <div>SignInPage</div>;
-};
-
-export default AuthHOC(SignInPage, false);
