--- src/page/SignUpPage.tsx
@@ -1,8 +1,0 @@
-import React from "react";
-import AuthHOC from "../components/HOC/AuthHOC";
-
-const SignUpPage = () => {
-  return <div>SignUpPage</div>;
-};
-
-export default AuthHOC(SignUpPage, false);
