+++ src/pages/SignInPage.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const SignInPage = () => {
+  return <div>SignInPage</div>;
+};
+
+export default SignInPage;
