+++ src/pages/SignUpPage.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const SignUpPage = () => {
+  return <div>SignUpPage</div>;
+};
+
+export default SignUpPage;
