+++ src/pages/TodoPage.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const TodoPage = () => {
+  return <div>TodoPage</div>;
+};
+
+export default TodoPage;
